Add two changes to G2P_FB118 material properties table
The first change has to do with the refractive index. As of now, it can be set (to a constant) via an input parameter. The second change has to do with the absorption length spectra, namely the WLS absorption length and the attenuation length (bulk absorption). Different spectra can be chosen up to the value of an input parameter, which reflects the WLS dye concentration in milligrams of WLS dye per PMMA kilogram.