Fix matching-optical-surfaces bug by introducing SiPMs protrusion
When tunneling probability is not null, the WLS plate is wrapped by G4LogicalSkinSurface. If blocks in between SiPMs are added, and the gap in between the WLS plate and the SiPMs surface is set to zero, then the logical skin surface of the WLS plate will match (in space) that of the SiPM board. This matching resulted in a bug due to which SiPMs were absorbed at the WLS plate side faces.