Fix compile-time error
As of the last rebase of this branch onto next-exp/master, SpherePointSampler::GenerateVertex does not take an string anymore, but a member of an enum which is defined within RandomUtils.h. As a consequence, now LArSphere::GenerateVertex, which still takes an string, should translate it into a member of the cited enum.