Change the name of specularspikeVIKUITI() to Vikuiti() and propagate it
This G4MaterialPropertiesTable is declared in OpticalMaterialProperties.h and defined in OpticalMaterialProperties.cc . This change has been propagated to every file where it is used. In XArapuca.cc, XArapucaDS.cc and XArapucaRods.cc, some changes need to be done when a new parameter to control de reflectivity type of vikuiti is added to the Vikuiti() G4MaterialPropertiesTable.