Changes to the absorption lengths of G2P_FB118 G4MaterialPropertiesT.
Before, the WLS-absorption length was only defined in the optical wavelength range of [330, 430] nm. Now it is defined in the [330, 600] nm range. Also, before, the attenuation (bulk absorption) length was defined in the following way: the measured pure-PMMA attenuation length in the [335, 430] nm range, and the measured dyed-PMMA absorption length in the [430, 600] nm range. The last contribution depends on the concentration. In this way, the set of defined attenuation and WLS-absorption lengths depended both on the cromophore concentration. Now, the implemented attenuation length is simply the measured attenuation length for the pure-PMMA, i.e. it is not dependent on the cromophore concentration anymore. The implemented data is raw in the [280, 550] nm range, but it is an extrapolation in the [550, 600] nm range though, because the measurement is too noisy in the >550 nm range.